requirements.txt
# ---------------------------------------- Core Libraries --------------------------------------------------------------
django==5.2.7
django-environ==0.12.0
django-model-utils==5.0.0
django-filter==25.2
Pillow==12.0.0
psycopg2-binary==2.9.11
argon2-cffi==25.1.0
gunicorn==23.0.0
phonenumbers==9.0.17
celery==5.5.3
ephem==4.2
pygments==2.19.2
whitenoise==6.11.0
redis==7.0.1


# ---------------------------------------- Django Packages -------------------------------------------------------------
djangorestframework==3.16.1
django-cors-headers==4.9.0
drf-spectacular==0.29.0
drf-standardized-errors[openapi]==0.15.0
djangorestframework-simplejwt==5.5.1
django-phonenumber-field==8.3.0
django-timezone-field==7.1
django-cleanup==9.0.0
django-unfold==1.0.0
django-import-export==4.3.13
django-anymail[resend]==13.1
django-storages[boto3,azure]==1.14.6
django_celery_beat==2.8.1
django_celery_results==2.6.0

# ---------------------------------------- Open Telemetry --------------------------------------------------------------
opentelemetry-exporter-otlp==1.38.0
opentelemetry-distro==0.59b0
# To find out what are the open telemetry packages that can be used: opentelemetry-bootstrap
# Following are the ones that were demed necessary for now.
opentelemetry-instrumentation-logging==0.59b0
opentelemetry-instrumentation-sqlite3==0.59b0
opentelemetry-instrumentation-threading==0.59b0
opentelemetry-instrumentation-celery==0.59b0
opentelemetry-instrumentation-django==0.59b0
opentelemetry-instrumentation-psycopg2==0.59b0
opentelemetry-instrumentation-requests==0.59b0
opentelemetry-instrumentation-redis==0.59b0
opentelemetry.instrumentation.urllib3==0.59b0
opentelemetry-instrumentation-urllib==0.59b0

# ---------------------------------------- Testing and Code Quality ----------------------------------------------------
coverage==7.11.0
pre-commit==4.3.0
mypy==1.18.2
djangorestframework-stubs==3.16.5
django-stubs==5.2.7
types-python-dateutil==2.9.0.20251008
types-requests==2.32.4.20250913

# ---------------------------------------- Debugging and Development Tools ---------------------------------------------
django-debug-toolbar==6.1.0
django-extensions==4.1
django-structlog==10.0.0
setuptools==80.9.0
django-zeal==2.0.4
django-silk==5.4.3
# ---------------------------------------- Package Managment Tools -----------------------------------------------------

# To upgrade pip packages: pip-upgrade
pip-upgrader==1.4.15


# Lanchain
langgraph==1.0.4
langchain-ollama==1.0.1
langchain-google-genai==4.1.1

# Presidio
presidio_analyzer==2.2.360
presidio_anonymizer==2.2.360
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl

# PDF text extractor
PyMuPDF==1.26.6
